+++ backend/srcs/src/channel/dto/index.ts
@@ -3,3 +3,5 @@ export * from './createChannelMessage.dto';
 export * from './createChannelPass.dto';
 export * from './joinChannel.dto';
 export * from './muteChannelUser.dto';
+export * from './responseChannel.dto';
+export * from './responseChannelUser.dto';