+++ frontend/srcs/src/app/services/index.ts
@@ -1,3 +1,4 @@
 export * from './api.service';
 export * from './auth.service';
 export * from './chat.service';
+export * from './channel.service';